services/cloud-connector-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
boto3>=1.34.0
azure-identity>=1.15.0
azure-mgmt-monitor>=6.0.0
azure-mgmt-compute>=30.0.0
azure-mgmt-resource>=23.0.0
google-cloud-monitoring>=2.0.0
google-cloud-resource-manager>=1.0.0
google-cloud-compute>=1.0.0
cryptography>=42.0.0
Django>=4.2
psycopg2-binary>=2.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
